Restructure Anvyl package into standard Python package layout
