+++ PACKAGE_RESTRUCTURE_SUMMARY.md
@@ -1,0 +1,92 @@
+# Anvyl Package Restructure Summary
+
+## Completed Restructuring
+
+The Anvyl project has been successfully restructured to ensure all anvyl-specific files are contained within the `anvyl` package directory. This means when installed via pip, all anvyl files will be properly contained within the `anvyl` directory in site-packages.
+
+## New Package Structure
+
+```
+anvyl/
+├── __init__.py                    # Main package init with version info
+├── cli.py                         # CLI entry point (renamed from anvyl_cli.py)
+├── grpc_server.py                # gRPC server (renamed from anvyl_grpc_server.py)
+├── proto_utils.py                # Protocol buffer utilities (moved from root)
+├── generate_protos.py            # Proto generation script (moved from root)
+├── sdk/
+│   ├── __init__.py               # SDK subpackage
+│   └── client.py                 # Anvyl client (moved from anvyl_sdk/)
+├── database/
+│   ├── __init__.py               # Database subpackage
+│   └── models.py                 # Database models (moved from database/)
+├── protos/
+│   └── anvyl.proto               # Protocol buffer definitions (moved from protos/)
+└── generated/
+    └── __init__.py               # Generated protobuf files directory
+```
+
+## Changes Made
+
+### File Movements
+- `anvyl_cli.py` → `anvyl/cli.py`
+- `anvyl_grpc_server.py` → `anvyl/grpc_server.py`
+- `proto_utils.py` → `anvyl/proto_utils.py`
+- `generate_protos.py` → `anvyl/generate_protos.py`
+- `anvyl_sdk/` contents → `anvyl/sdk/`
+- `database/` contents → `anvyl/database/`
+- `protos/` contents → `anvyl/protos/`
+
+### Configuration Updates
+
+#### pyproject.toml
+- Updated CLI entry point: `anvyl = "anvyl.cli:app"`
+- Updated package discovery: `include = ["anvyl*"]`
+- Updated exclusions to prevent UI/scripts from being installed
+- Updated tool configurations for the new paths
+- Removed `py-modules` (no longer needed)
+- Added proper `package-data` configuration
+
+#### MANIFEST.in
+- Updated to include `anvyl` package recursively
+- Removed individual file includes
+- Properly excludes UI, scripts, and other non-package files
+
+### Import Updates
+- Fixed relative imports in all moved files
+- Updated CLI to use `from .sdk import AnvylClient`
+- Updated gRPC server to use `from .generated import anvyl_pb2`
+- Updated proto_utils paths for new structure
+- Added missing methods to SDK client for CLI compatibility
+
+### Package Structure
+- Added proper `__init__.py` files for all subpackages
+- Main package exports `client`, `models`, and `__version__`
+- Maintained backward compatibility where possible
+
+## Installation Benefits
+
+After this restructuring:
+
+1. **Clean site-packages**: All anvyl files will be contained in `anvyl/` directory
+2. **Proper Python package**: Standard Python package structure with subpackages
+3. **Better imports**: Clear import hierarchy (`from anvyl.sdk import client`)
+4. **No pollution**: UI, scripts, and docs won't be installed with the package
+5. **Maintainable**: Standard structure makes the codebase easier to maintain
+
+## CLI Entry Point
+
+The CLI is still accessible via:
+```bash
+anvyl --help
+```
+
+The entry point now correctly points to `anvyl.cli:app` in the package structure.
+
+## Development Usage
+
+For development, the package can be installed in editable mode:
+```bash
+pip install -e .
+```
+
+All imports will work correctly with the new structure.
