+++ pyproject.toml
@@ -87,7 +87,7 @@ all = [
 ]
 
 [project.scripts]
-anvyl = "anvyl_cli:app"
+anvyl = "anvyl.cli:app"
 
 [project.urls]
 Homepage = "https://github.com/kessler-frost/anvyl"
@@ -102,7 +102,7 @@ target-version = ['py312']
 include = '\.pyi?$'
 extend-exclude = '''
 /(
-  generated/.*\.py$
+  anvyl/generated/.*\.py$
 )/
 '''
 
@@ -114,33 +114,33 @@ force_grid_wrap = 0
 use_parentheses = true
 ensure_newline_before_comments = true
 line_length = 88
-skip_glob = ["generated/*.py"]
+skip_glob = ["anvyl/generated/*.py"]
 
 [tool.mypy]
 python_version = "3.12"
 warn_return_any = true
 warn_unused_configs = true
 disallow_untyped_defs = true
-exclude = ['generated/.*\.py$']
+exclude = ['anvyl/generated/.*\.py$']
 
 [[tool.mypy.overrides]]
 module = [
     "docker.*",
     "grpc.*",
-    "generated.*",
+    "anvyl.generated.*",
 ]
 ignore_missing_imports = true
 
 [tool.setuptools]
-py-modules = ["anvyl_cli", "anvyl_grpc_server", "proto_utils"]
 include-package-data = true
 
 [tool.setuptools.packages.find]
 where = ["."]
-include = ["anvyl_sdk*", "database*", "generated*"]
-exclude = ["tests*"]
+include = ["anvyl*"]
+exclude = ["tests*", "ui*", "scripts*", "docs*"]
 
 [tool.setuptools.package-data]
-"*" = ["*.proto"]
+"anvyl" = ["protos/*.proto"]
+"anvyl.generated" = ["*.py"]
 
 # Note: Protobuf files are auto-generated when the package is first imported
